+++ lizard_datasource/augmented_datasource.py
@@ -28,32 +28,36 @@ class AugmentedDataSource(datasource.DataSource):
     """By default this just forwards most important methods to the
     datasource that's being augmented. Before returning some results
     may or may not be, well, augmented."""
-    def __init__(self, augmented_model_instance):
-        self.augmented_model_instance = augmented_model_instance
+
+    def __init__(self, config_object):
+        """Initialize the object.
+
+        The config_object is an instance of the AugmentedDataSource
+        model. It is used to store configuration for this augmented
+        datasource in."""
+
+        self.config_object = config_object
 
     @property
-    def augmented_source(self):
-        if not hasattr(self, '_augmented_source'):
-            logger.debug("PING")
-            try:
-                self._augmented_source = datasource.get_datasource_by_model(
-                    self.augmented_model_instance.augmented_source,
-                    exclude=self)
-            except Exception, e:
-                logger.debug("But no PONG: {0}.".format(e))
-            logger.debug("The chosen augmented source is: {0}."
-                         .format(self._augmented_source))
-        return self._augmented_source
+    def original_datasource(self):
+        """Return the datasource object that is augmented by this
+        AugmentedDataSource object."""
+
+        if not hasattr(self, '_original_datasource'):
+            self._original_datasource = datasource.get_datasource_by_model(
+                self.config_object.augmented_source,
+                exclude=self)
+
+        return self._original_datasource
 
     @property
     def PROPERTIES(self):
-        logger.debug("I'm in PROPERTIES and my id is {0}.".format(id(self)))
-        return self.augmented_source.PROPERTIES
+        return self.original_datasource.PROPERTIES
 
     @property
     def identifier(self):
-        return "augmented_{0}".format(
-            self.augmented_model_instance.id)
+        return "augmented_{0}_{1}".format(
+            self.config_object.id, self.config_object.name)
 
     @property
     def originating_app(self):
@@ -61,92 +65,137 @@ class AugmentedDataSource(datasource.DataSource):
 
     def set_choices_made(self, choices_made):
         self._choices_made = choices_made
-        self.augmented_source.set_choices_made(choices_made)
+        self.original_datasource.set_choices_made(choices_made)
 
     def criteria(self):
-        return self.augmented_source.criteria()
+        return self.original_datasource.criteria()
 
     def options_for_criterion(self, criterion):
-        return self.augmented_source.options_for_criterion(criterion)
+        return self.original_datasource.options_for_criterion(criterion)
 
     def chooseable_criteria(self):
-        return self.augmented_source.chooseable_criteria()
+        """These are the chooseable criteria from the augmented
+        source, minus the layers that are hidden because they are used
+        for stuff like colors and percentiles."""
+
+        criteria = self.original_datasource.chooseable_criteria()
+
+        colorfroms = models.ColorFromLatestValue.objects.filter(
+            augmented_source=self.config_object,
+            hide_from_layer=True)
+        percentiles = models.PercentileLayer.objects.filter(
+            augmented_source=self.config_object,
+            hide_from_layer=True)
+
+        forbidden_choices_json = [
+            colorfrom.layer_to_get_color_from.choices_made
+            for colorfrom in colorfroms] + [
+            percentile.layer_to_get_percentile_from.choices_made
+            for percentile in percentiles]
+
+        clean_criteria = []
+
+        my_choices = self._choices_made
+
+        while criteria:
+            retry_criteria = []
+            for chooseable in criteria:
+                criterion = chooseable['criterion']
+                options = chooseable['options']
+
+                for option in options.iter_options():
+                    new_choices = my_choices.add_criterion_option(
+                        criterion, option)
+                    if new_choices.json() in forbidden_choices_json:
+                        # Alas, we can't use these options as is,
+                        # there is a forbidden option in them.
+                        if len(options) > 1:
+                            # There is more than one option. Retry this
+                            # criterion with one option removed.
+                            retry_criteria.append({
+                                    'criterion': criterion,
+                                    'options': options.minus(option)
+                                    })
+                        break
+                else:
+                    # Not forbidden, we can use these options!
+                    clean_criteria.append(chooseable)
+
+            # Restart the loop with the retry criteria, if any
+            criteria = retry_criteria
+
+        return clean_criteria
 
     def is_applicable(self, choices_made):
-        return self.augmented_source.is_applicable(choices_made)
+        return self.original_datasource.is_applicable(choices_made)
 
     def is_drawable(self, choices_made):
-        return self.augmented_source.is_drawable(choices_made)
+        return self.original_datasource.is_drawable(choices_made)
 
     def locations(self, bare=False):
-        locations = list(self.augmented_source.locations(bare=bare))
+        locations = list(self.original_datasource.locations(bare=bare))
 
         if bare:
             for location in locations:
                 yield location
             return
 
-        datasource_layer = None
-        if self._choices_made.get('parameter', None) == 'GW.meting':
-            klasse = self._choices_made.add(
-                'parameter', 'klasse')
+        cached_values = dict()
+        try:
+            colorfrom = models.ColorFromLatestValue.objects.get(
+                layer_to_add_color_to=self.datasource_layer)
 
-            old_choices = self._choices_made
-            self._choices_made = klasse
-            datasource_layer = self.datasource_layer
-            self._choices_made = old_choices
-
-            caches = dict()
-            if datasource_layer:
-                for cache in models.DatasourceCache.objects.filter(
-                    datasource_layer=datasource_layer):
-                    caches[cache.locationid] = cache.value
-            colormap = models.ColorMap.objects.get(pk=1)
+            # If there is a colorfrom, use it
+            colormap = colorfrom.colormap
+            if colorfrom.layer_to_get_color_from:
+                for cached_value in models.DatasourceCache.objects.filter(
+                    datasource_layer=colorfrom.layer_to_get_color_from):
+                    cached_values[cached_value.locationid] = (
+                        cached_value.value)
+        except models.ColorFromLatestValue.DoesNotExist:
+            # No colouring.
+            pass
 
         for location in locations:
             color = "888888"  # Default is gray
-            if datasource_layer:
-                if location.identifier in caches:
-                    value = caches[location.identifier]
-                    color = colormap.color_for(value)
-                    if color.startswith("#"):
-                        color = color[1:]
+            if location.identifier in cached_values:
+                value = cached_values[location.identifier]
+                color = colormap.color_for(value)
+                if color.startswith("#"):
+                    color = color[1:]
 
             location.color = color
             yield location
 
     def timeseries(self, location_id, start_datetime=None, end_datetime=None):
-        return self.augmented_source.timeseries(
+        return self.original_datasource.timeseries(
             location_id, start_datetime, end_datetime)
 
     def has_percentiles(self):
-        if self._choices_made.get('parameter', None) == 'GW.meting':
-            logger.debug("Parameter is GW.meting!")
-            return True
-        return False
+        return models.PercentileLayer.objects.filter(
+            layer_to_add_percentile_to=self.datasource_layer
+            ).exists()
 
     def percentiles(self, location_id, start_datetime=None, end_datetime=None):
         percentiles = {}
-        for percentile, parameter in ((10, 'GW.meting.10'),
-                                      (25, 'GW.meting.25'),
-                                      (75, 'GW.meting.75'),
-                                      (90, 'GW.meting.90')):
-            self.augmented_source.set_choices_made(
-                self._choices_made.add('parameter', parameter))
-            percentiles[percentile] = self.augmented_source.timeseries(
+        for percentile_layer in models.PercentileLayer.objects.filter(
+            layer_to_add_percentile_to=self.datasource_layer):
+            source = datasource.get_datasource_by_layer(
+                percentile_layer.layer_to_get_percentile_from)
+
+            percentiles[percentile_layer.percentile] = source.timeseries(
                 location_id, start_datetime, end_datetime).data()
-        self.augmented_source.set_choices_made(self._choices_made)
-        logger.debug("Percentiles: {0}".format(percentiles))
+
         return percentiles
 
 
 def factory():
-    """Returns an AugmentedDataSource object for each
+    """Return an AugmentedDataSource object for each
     AugmentedDataSource model instance."""
 
     try:
-        return [AugmentedDataSource(source)
-                for source in models.AugmentedDataSource.objects.all()]
+        return [AugmentedDataSource(config_object)
+                for config_object in models.AugmentedDataSource.objects.all()]
     except Exception, e:
         logger.debug(
             "Exception in augmented datasource factory: {0}".format(e))