+++ CHANGES.rst
@@ -5,7 +5,20 @@ Changelog of lizard-datasource
 0.3 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Removed hardcoding of options.
+
+- The cache script now creates all layer objects, but only downloads
+  values for layers of which the latest value is used by something
+  (coloring).
+
+- Coloring is done using the configuration in the ColorFromLatestValue
+  model.
+
+- Percentile bars are done using the configuration in the
+  PercentileLayer model.
+
+- Layers that are used for augmenting other layers (color layers,
+  percentile layers) can now be hidden.
 
 
 0.2 (2012-12-20)